Removed support for Fahrenheit
The functions to convert between Celsius and Fahrenheit are kept.
However, specific functionality that reports temperatures in Fahrenheit
are removed. This means the boolean parameter to `readTemperature()`
and the `computeHeatIndexF()`. The library reports all temperatures in
Celsius.
The user is required to do their own conversion to Fahrenheit. As
mentioned conversion functions to do that remain in the library.